refact: 뮤직 아이템 컴포넌트를 분리시킴
