+++ src/components/musicList/MusicItem.module.scss
@@ -1,0 +1,65 @@
+@import '/src/global.scss';
+
+.wrapper{ 
+    display: flex;
+    align-items: center;
+    width: 100%;
+}
+.wrapper--selected{
+    background-color: rgba(255, 255, 255, 0.1);
+}
+.grid-container {
+    width: 100%;
+    padding: 5px 10px;
+
+    display: grid;    
+    grid-template-columns: 45px auto 50px;
+    grid-gap: 10px 10px;    
+    justify-items: stretch;
+    align-items: center;
+    &:hover .overlay{
+        visibility: initial;
+    }
+}
+.grid-item{
+    position: relative;
+    overflow: hidden;
+    text-align: center;
+}
+
+.grid-item--fit{
+    width: 100%;
+    height: 100%;
+}
+.grid-item__text{
+    width: 100%;
+    padding-left: 10px;
+    text-align: left;
+    font-size: 16px;
+    line-height: 1.3;
+    font-weight: 500;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
+}
+.grid-item__text--light{
+    font-weight: 400;
+    color:$font-secondary-color
+}
+.thumbnail{
+    object-fit: cover;
+    width: 100%;
+    height: 100%;
+}
+.overlay{
+    position: absolute;
+    top: 0px;
+    width: 100%;
+    height: 100%;
+    padding: 4px;
+    background-color: rgba($color: #000000, $alpha: 0.4);
+    visibility: hidden;
+}
+.overlay--show{
+    visibility: initial;
+}
