+++ src/components/musicList/MusicList.module.scss
@@ -5,31 +5,6 @@
     overflow: auto;
     @include append_scrollbar;
 }
-.drag_handle{
-    width: 32px;
-    padding-left: 10px;
-}
-.grid-item__icon{
-    width: 32px;
-    height:32px;
-    object-fit: contain;
-    margin: 0 16px 0 0;
-}
-.grid-item__text{
-    width: 100%;
-    padding-left: 10px;
-    text-align: left;
-    font-size: 16px;
-    line-height: 1.3;
-    font-weight: 500;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-}
-.grid-item__text--light{
-    font-weight: 400;
-    color:$font-secondary-color
-}
 .wrapper{ 
     display: flex;
     align-items: center;
@@ -38,44 +13,8 @@
 .wrapper--selected{
     background-color: rgba(255, 255, 255, 0.1);
 }
-.grid-container {
-    width: 100%;
-    padding: 5px 10px;
 
-    display: grid;    
-    grid-template-columns: 45px auto 50px;
-    grid-gap: 10px 10px;    
-    justify-items: stretch;
-    align-items: center;
-    &:hover .overlay{
-        visibility: initial;
-    }
-}
-.grid-item{
-    position: relative;
-    overflow: hidden;
-    text-align: center;
-    
-}
-.grid-item--fit{
-    width: 100%;
-    height: 100%;
-}
-.overlay{
-    position: absolute;
-    top: 0px;
-    width: 100%;
-    height: 100%;
-    padding: 4px;
-    background-color: rgba($color: #000000, $alpha: 0.4);
-    visibility: hidden;
-}
-.overlay--show{
-    visibility: initial;
-}
-
-.thumbnail{
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
+.drag_handle{
+    width: 32px;
+    padding-left: 10px;
 }
